:recycle: CHORE : 충돌 수정 및 Message 엔티티 수정
