+++ src/main/java/com/hipet/domain/User/entity/User.java
@@ -31,9 +31,6 @@ public class User extends BaseEntity {
     @Column(name="USER_NAME")
     private String userName; //사용자 이름
 
-    @Column(name="ADDRESS")
-    private String address; //주소 -> 범위는 미정..
-
     @Column(name="PROFILE_INFO")
     private String profileInfo; //사용자 소개
 
@@ -56,7 +53,7 @@ public class User extends BaseEntity {
     public void changeUserName(String newUserName) {this.userName = newUserName;}
 
     //사용자 주소 수정 메소드
-    public void changeAddress(String newAddress) {this.address = newAddress;}
+    public void changeAddress(String newRegion) {this.region = Region.valueOf(newRegion);}
 
     //사용자 소개 수정 메소드
     public void changeProfileInfo(String newProfileInfo) {this.profileInfo = newProfileInfo;}
